🚚 Move CardWithIcon to shared folder
