+++ src/components/BabyInfo/BabyAvatar.tsx
@@ -2,8 +2,7 @@ import { Avatar, AvatarProps } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import { baby } from "../../assets";
 import { calculateAge } from "../../utils";
-import { CardWithIcon } from "../CardWithIcon";
-import { Typography } from "../Shared";
+import { CardWithIcon, Typography } from "../Shared";
 
 export type BabyAvatarProps = AvatarProps & {
   name: string;
