+++ src/components/Shared/index.tsx
@@ -1,1 +1,2 @@
+export * from "./CardWithIcon";
 export * from "./MuiWrap";
